fix:各種リファクタリング及び微修正.
