+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Rooms from "./components/rooms/Rooms";
 import Calendar from "./components/schedule/calendar/Calendar";
 
-export default function Home() {
+export default async function Home() {
   return (
     <main>
       <Rooms />
